coprophage/requirements.txt
kafka-python==2.0.2
psycopg2==2.9.9
cassandra-driver==3.28.0
